# perf: set swc minifier.compress.passes to 1