--- a/lib/minifiers/swc.ts
+++ b/lib/minifiers/swc.ts
@@ -4,7 +4,7 @@ import { minifier } from '../types';
 export default minifier(async ({ code }) => {
 	const minified = await minify(code, {
 		compress: {
-			// passes: 1,
+			passes: 1,
 		},
 		mangle: true,
 		sourceMap: false,