@privatenumber 2022-12-07
I think it's better to stick with default configuration to demonstrate what users should expect out of the box.

If a specific configuration yields drastically different results, then maybe it should be added as a separate minifier so they can be compared. In this case, I think the differences aren't big enough.