@kdy1 2022-12-07
Same opinion as @privatenumber 
The default value is carefully determined.